Created separate python scripts to save individual concrete implementations of 'ExitStruct' abstract class.
