+++ src/strat_backtest/base/exit_struct.py
@@ -6,13 +6,16 @@ from abc import ABC, abstractmethod
 from collections import deque
 from datetime import datetime
 from decimal import Decimal
-from typing import Any
+from typing import TYPE_CHECKING
 
 from pydantic import ValidationError
 
-from strat_backtest.utils import get_net_pos
+from strat_backtest.utils import ClosedPositionResult, OpenTrades, get_net_pos
 
-from .stock_trade import StockTrade
+if TYPE_CHECKING:
+    from strat_backtest.utils import CompletedTrades
+
+    from .stock_trade import StockTrade
 
 
 class ExitStruct(ABC):
@@ -32,25 +35,25 @@ class ExitStruct(ABC):
     @abstractmethod
     def close_pos(
         self,
-        open_trades: deque[StockTrade],
+        open_trades: OpenTrades,
         dt: datetime,
-        exit_price: float,
-    ) -> tuple[deque[StockTrade], list[dict[str, Any]]]:
+        exit_price: Decimal,
+    ) -> ClosedPositionResult:
         """Update existing StockTrade objects (still open); and remove completed
         StockTrade objects in 'open_trades'.
 
         Args:
-            open_trades (deque[StockTrade]):
+            open_trades (OpenTrades):
                 Deque list of StockTrade pydantic object to record open trades.
             dt (datetime):
                 Trade datetime object.
-            exit_price (float):
+            exit_price (Decimal):
                 Exit price of stock ticker.
 
         Returns:
-            open_trades (deque[StockTrade]):
+            open_trades (OpenTrades):
                 Updated deque list of 'StockTrade' objects.
-            completed_trades (list[dict[str, Any]]):
+            completed_trades (CompletedTrades):
                 List of dictionary containing required fields to generate DataFrame.
         """
 
@@ -58,11 +61,11 @@ class ExitStruct(ABC):
 
     def _update_pos(
         self,
-        trade: StockTrade,
+        trade: "StockTrade",
         dt: datetime,
         exit_price: float,
         exit_lots: int | None = None,
-    ) -> StockTrade:
+    ) -> "StockTrade":
         """Update existing StockTrade objects (still open).
 
         Args:
@@ -98,7 +101,7 @@ class ExitStruct(ABC):
             print(f"Validation Error : {e}")
             return trade
 
-    def _validate_completed_trades(self, stock_trade: StockTrade) -> bool:
+    def _validate_completed_trades(self, stock_trade: "StockTrade") -> bool:
         """Validate whether StockTrade object is properly updated with no null
         values."""
 
@@ -111,392 +114,3 @@ class ExitStruct(ABC):
         is_lots_matched = stock_trade.entry_lots == stock_trade.exit_lots
 
         return is_no_null_field and is_lots_matched
-
-
-class FIFOExit(ExitStruct):
-    """Take profit for earliest open trade. For example:
-
-    - Long stock at 50 (5 lots) -> 60 (3 lots) -> 70 2 lots).
-    - Signal to take profit -> Sell 5 lots bought at 50 (i.e. FIFO).
-    """
-
-    def close_pos(
-        self,
-        open_trades: deque[StockTrade],
-        dt: datetime,
-        exit_price: float,
-    ) -> tuple[deque[StockTrade], list[dict[str, Any]]]:
-        """Update existing StockTrade objects (still open); and remove completed
-        StockTrade objects in 'open_trades'.
-
-        Args:
-            open_trades (deque[StockTrade]):
-                Deque list of StockTrade pydantic object to record open trades.
-            dt (datetime):
-                Trade datetime object.
-            exit_price (float):
-                Exit price of stock ticker.
-
-        Returns:
-            open_trades (deque[StockTrade]):
-                Updated deque list of 'StockTrade' objects.
-            completed_trades (list[dict[str, Any]]):
-                List of dictionary containing required fields to generate DataFrame.
-        """
-
-        if len(open_trades) == 0:
-            # No open trades to close
-            return open_trades, []
-
-        completed_trades = []
-        earliest_trade = open_trades[0]
-
-        # Update earliest StockTrade object
-        earliest_trade = self._update_pos(earliest_trade, dt, exit_price)
-
-        # Convert StockTrade to dictionary only if all fields are populated
-        # i.e. trade completed.
-        if self._validate_completed_trades(earliest_trade):
-            # Remove earliest StockTrade object since it is completed
-            open_trades.popleft()
-
-            completed_trades.append(earliest_trade.model_dump())
-
-        return open_trades, completed_trades
-
-
-class LIFOExit(ExitStruct):
-    """Take profit for latest open trade. For example:
-
-    - Long stock at 50 (5 lots) -> 60 (3 lots) -> 70 (2 lots).
-    - Signal to take profit -> Sell 2 lots bought at 70 (i.e. FIFO).
-    """
-
-    def close_pos(
-        self,
-        open_trades: deque[StockTrade],
-        dt: datetime,
-        exit_price: float,
-    ) -> tuple[deque[StockTrade], list[dict[str, Any]]]:
-        """Update existing StockTrade objects (still open); and remove completed
-        StockTrade objects in 'open_trades'.
-
-        Args:
-            open_trades (deque[StockTrade]):
-                Deque list of StockTrade pydantic object to record open trades.
-            dt (datetime):
-                Trade datetime object.
-            exit_price (float):
-                Exit price of stock ticker.
-
-        Returns:
-            open_trades (deque[StockTrade]):
-                Updated deque list of 'StockTrade' objects.
-            completed_trades (list[dict[str, Any]]):
-                List of dictionary containing required fields to generate DataFrame.
-        """
-
-        if len(open_trades) == 0:
-            # No open trades to close
-            return open_trades, []
-
-        completed_trades = []
-        latest_trades = open_trades[-1]
-
-        # Update earliest StockTrade object
-        latest_trades = self._update_pos(latest_trades, dt, exit_price)
-
-        # Convert StockTrade to dictionary only if all fields are populated
-        # i.e. trade completed.
-        if self._validate_completed_trades(latest_trades):
-            # Remove latest StockTrade object from 'open_trades'
-            open_trades.pop()
-
-            completed_trades.append(latest_trades.model_dump())
-
-        return open_trades, completed_trades
-
-
-class HalfFIFOExit(ExitStruct):
-    """keep taking profit by exiting half of earliest positions. For example:
-
-    - Long stock at 50 (5 lots) -> 60 (3 lots) -> 70 (2 lots).
-    - Signal to take profit -> Sell 5 lots (50% of total 10 lots) bought at 50
-    i.e. left 60 (3 lots), 70 (2 lots)
-    - Signal to take profit again -> sell 3 lots (50% of total 5 lots) bought at 60
-    i.e. left 70 (2 lots).
-    """
-
-    def close_pos(
-        self,
-        open_trades: deque[StockTrade],
-        dt: datetime,
-        exit_price: float,
-    ) -> tuple[deque[StockTrade], list[dict[str, Any]]]:
-        """Update existing StockTrade objects (still open); and remove completed
-        StockTrade objects in 'open_trades'.
-
-        Args:
-            open_trades (deque[StockTrade]):
-                Deque list of StockTrade pydantic object to record open trades.
-            dt (datetime):
-                Trade datetime object.
-            exit_price (float):
-                Exit price of stock ticker.
-
-        Returns:
-            new_open_trades (deque[StockTrade]):
-                Updated deque list of 'StockTrade' objects.
-            completed_trades (list[dict[str, Any]]):
-                List of dictionary containing required fields to generate DataFrame.
-        """
-
-        completed_trades = []
-        new_open_trades = deque()
-
-        if len(open_trades) == 0:
-            # No open trades to close
-            return open_trades, []
-
-        # Get net position and half of net position from 'open_trades'
-        net_pos = get_net_pos(open_trades)
-        half_pos = math.ceil(abs(net_pos) / 2)
-
-        for trade in open_trades:
-            initial_exit_lots = trade.exit_lots
-
-            # Update trade only if haven't reach half of net position
-            if half_pos > 0:
-                # Determine quantity to close based on 'half_pos'
-                lots_to_exit = min(half_pos, trade.entry_lots - initial_exit_lots)
-
-                # Update StockTrade objects with exit info
-                trade = self._update_pos(
-                    trade, dt, exit_price, initial_exit_lots + lots_to_exit
-                )
-
-                # Break loop if trade is not updated properly i.e.
-                # trade.exit_lots = initial_exit_lots
-                if trade.exit_lots == initial_exit_lots:
-                    return open_trades, []
-
-                # Only update 'new_open_trades' if trades are still partially closed
-                if not self._validate_completed_trades(trade):
-                    new_open_trades.append(trade)
-
-                # Create completed trades using 'lots_to_exit'
-                completed_trades.append(self._gen_completed_trade(trade, lots_to_exit))
-
-                # Update remaining positions required to be closed and net position
-                half_pos -= lots_to_exit
-
-            # trade not updated
-            else:
-                new_open_trades.append(trade)
-
-        return new_open_trades, completed_trades
-
-    def _gen_completed_trade(
-        self, trade: StockTrade, lots_to_exit: Decimal
-    ) -> dict[str, Any]:
-        """Generate StockTrade object with completed trade from 'StockTrade'
-        and convert to dictionary."""
-
-        # Create a shallow copy of the updated trade
-        completed_trade = trade.model_copy()
-
-        # Update the 'entry_lots' and 'exit_lots' to be same as 'lots_to_exit'
-        completed_trade.entry_lots = lots_to_exit
-        completed_trade.exit_lots = lots_to_exit
-
-        if not self._validate_completed_trades(completed_trade):
-            raise ValueError("Completed trades not properly closed.")
-
-        return completed_trade.model_dump()
-
-
-class HalfLIFOExit(ExitStruct):
-    """keep taking profit by exiting half of latest positions . For example:
-
-    - Long stock at 50 (5 lots) -> 60 (3 lots) -> 70 (2 lots).
-    - Signal to take profit -> Sell 2 lots bought at 70 and 3 lots bought at 60
-    i.e. left 50 (5 lots)
-    - Signal to take profit again -> sell 3 lots bought at 50
-    i.e. left 50 (2 lots).
-    """
-
-    def close_pos(
-        self,
-        open_trades: deque[StockTrade],
-        dt: datetime,
-        exit_price: float,
-    ) -> tuple[deque[StockTrade], list[dict[str, Any]]]:
-        """Update existing StockTrade objects (still open); and remove completed
-        StockTrade objects in 'open_trades'.
-
-        Args:
-            open_trades (deque[StockTrade]):
-                Deque list of StockTrade pydantic object to record open trades.
-            dt (datetime):
-                Trade datetime object.
-            exit_price (float):
-                Exit price of stock ticker.
-
-        Returns:
-            new_open_trades (deque[StockTrade]):
-                Updated deque list of 'StockTrade' objects.
-            completed_trades (list[dict[str, Any]]):
-                List of dictionary containing required fields to generate DataFrame.
-        """
-
-        completed_trades = []
-        new_open_trades = deque()
-
-        if len(open_trades) == 0:
-            # No open trades to close
-            return open_trades, []
-
-        # Reverse copy of 'self.open_trades'
-        open_trades_list = list(open_trades)
-        reversed_open_trades = open_trades_list[::-1]
-
-        # Get net position and half of net position from 'open_trades'
-        net_pos = get_net_pos(open_trades)
-        half_pos = math.ceil(abs(net_pos) / 2)
-
-        for trade in reversed_open_trades:
-            initial_exit_lots = trade.exit_lots
-
-            # Update trade only if haven't reach half of net position
-            if half_pos > 0:
-                # Determine quantity to close based on 'half_pos'
-                lots_to_exit = min(half_pos, trade.entry_lots - initial_exit_lots)
-
-                # Update StockTrade objects with exit info
-                trade = self._update_pos(
-                    trade, dt, exit_price, initial_exit_lots + lots_to_exit
-                )
-
-                # Break loop if trade is not updated properly i.e.
-                # trade.exit_lots = initial_exit_lots
-                if trade.exit_lots == initial_exit_lots:
-                    return open_trades, []
-
-                # Only update 'new_open_trades' if trades are still partially closed
-                if not self._validate_completed_trades(trade):
-                    new_open_trades.append(trade)
-
-                # Create completed trades using 'lots_to_exit'
-                completed_trades.append(self._gen_completed_trade(trade, lots_to_exit))
-
-                # Update remaining positions required to be closed and net position
-                half_pos -= lots_to_exit
-
-            # trade not updated
-            else:
-                new_open_trades.appendleft(trade)
-
-        return new_open_trades, completed_trades
-
-    def _gen_completed_trade(
-        self, trade: StockTrade, lots_to_exit: Decimal
-    ) -> dict[str, Any]:
-        """Generate StockTrade object with completed trade from 'StockTrade'
-        and convert to dictionary."""
-
-        # Create a shallow copy of the updated trade
-        completed_trade = trade.model_copy()
-
-        # Update the 'entry_lots' and 'exit_lots' to be same as 'lots_to_exit'
-        completed_trade.entry_lots = lots_to_exit
-        completed_trade.exit_lots = lots_to_exit
-
-        if not self._validate_completed_trades(completed_trade):
-            raise ValueError("Completed trades not properly closed.")
-
-        return completed_trade.model_dump()
-
-
-class TakeAllExit(ExitStruct):
-    """Exit all open positions at a loss."""
-
-    def close_pos(
-        self,
-        open_trades: deque[StockTrade],
-        dt: datetime,
-        exit_price: float,
-    ) -> tuple[deque[StockTrade], list[dict[str, Any]]]:
-        """Update existing StockTrade objects (still open); and remove completed
-        StockTrade objects in 'open_trades'.
-
-        Args:
-            open_trades (deque[StockTrade]):
-                Deque list of StockTrade pydantic object to record open trades.
-            dt (datetime):
-                Trade datetime object.
-            exit_price (float):
-                Exit price of stock ticker.
-
-        Returns:
-            open_trades (deque[StockTrade]):
-                Updated deque list of 'StockTrade' objects.
-            completed_trades (list[dict[str, Any]]):
-                List of dictionary containing required fields to generate DataFrame.
-        """
-
-        if len(open_trades) == 0:
-            # No open trades to close
-            return open_trades, []
-
-        completed_trades = []
-
-        for trade in open_trades:
-            initial_exit_lots = trade.exit_lots
-
-            # Update to close open position
-            trade = self._update_pos(trade, dt, exit_price)
-
-            # Break loop if trade is not updated properly i.e.
-            # trade.exit_lots = initial_exit_lots
-            if trade.exit_lots == initial_exit_lots:
-                # Return original 'open_trades' and empty completed trades list
-                # if trade is None i.e. ValidationError
-                return open_trades, []
-
-            if initial_exit_lots > 0:
-                # Number of lots to close partial open position
-                lots_to_exit = trade.entry_lots - initial_exit_lots
-
-                # Create completed trades using 'lots_to_exit'
-                completed_trades.append(self._gen_completed_trade(trade, lots_to_exit))
-
-            else:
-                # Generate completed trades for fully open position
-                if self._validate_completed_trades(trade):
-                    completed_trades.append(trade.model_dump())
-
-        if len(completed_trades) != len(open_trades):
-            raise ValueError("Open positions failed to close completely.")
-
-        # Reset open_trades
-        open_trades.clear()
-
-        return open_trades, completed_trades
-
-    def _gen_completed_trade(
-        self, trade: StockTrade, lots_to_exit: Decimal
-    ) -> dict[str, Any]:
-        """Generate StockTrade object with completed trade from 'StockTrade'
-        and convert to dictionary."""
-
-        # Create a shallow copy of the updated trade
-        completed_trade = trade.model_copy()
-
-        # Update the 'entry_lots' and 'exit_lots' to be same as 'lots_to_exit'
-        completed_trade.entry_lots = lots_to_exit
-        completed_trade.exit_lots = lots_to_exit
-
-        if not self._validate_completed_trades(completed_trade):
-            raise ValueError("Completed trades not properly closed.")
-
-        return completed_trade.model_dump()
