+++ src/strat_backtest/base/stock_trade.py
@@ -2,7 +2,6 @@
 
 from datetime import datetime
 from decimal import Decimal
-from typing import Any
 
 from pydantic import BaseModel, Field, computed_field, field_validator
